released section refactor
