+++ frontend/src/components/managerialConsole/ReleasedSection.tsx
@@ -1,14 +1,9 @@
-import React, { useState } from 'react';
 import { BaseSection } from '@components/core/BaseSection';
-import { useTranslation } from 'react-i18next';
-import moment from 'moment';
-import { useBaseUrl } from '@hooks/useBaseUrl';
 import { BlackLink } from '@components/core/BlackLink';
 import { UniversalMoment } from '@components/core/UniversalMoment';
+import { useBaseUrl } from '@hooks/useBaseUrl';
+import SearchIcon from '@mui/icons-material/Search';
 import {
-  MenuItem,
-  Select,
-  TextField,
   InputAdornment,
   Table,
   TableBody,
@@ -17,12 +12,11 @@ import {
   TableHead,
   TableRow,
   TableSortLabel,
-  SelectChangeEvent,
-  IconButton,
-  Box,
+  TextField,
 } from '@mui/material';
-import SearchIcon from '@mui/icons-material/Search';
-import { Close } from '@mui/icons-material';
+import React, { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { ProgramSelect, useSortAndFilter } from './useSortAndFilter';
 
 interface ReleasedSectionProps {
   releasedData: any;
@@ -34,20 +28,15 @@ export const ReleasedSection: React.FC<ReleasedSectionProps> = ({
   const { t } = useTranslation();
   const { businessArea } = useBaseUrl();
   const [searchText, setSearchText] = useState('');
-  const [sortField, setSortField] = useState(null);
-  const [sortDirection, setSortDirection] = useState<'asc' | 'desc'>('asc');
-  const [selectedProgram, setSelectedProgram] = useState('');
-
-  const handleSort = (field) => {
-    const newSortDirection =
-      sortField === field && sortDirection === 'asc' ? 'desc' : 'asc';
-    setSortField(field);
-    setSortDirection(newSortDirection);
-  };
-
-  const handleProgramChange = (event: SelectChangeEvent<string>) => {
-    setSelectedProgram(event.target.value);
-  };
+  const {
+    sortField,
+    sortDirection,
+    selectedProgram,
+    setSelectedProgram,
+    handleSort,
+    sortRows,
+    filterRows,
+  } = useSortAndFilter({ initialSortField: null, initialSortDirection: 'asc' });
 
   const programs = releasedData?.results?.reduce((acc, row) => {
     if (!acc.includes(row.program)) {
@@ -84,37 +73,13 @@ export const ReleasedSection: React.FC<ReleasedSectionProps> = ({
     },
   ];
 
-  const filteredRows =
-    releasedData?.results?.filter((row: any) =>
-      columns.some((column) => {
-        if (selectedProgram !== '' && row.program !== selectedProgram) {
-          return false;
-        }
-        if (column.field === 'program') {
-          return (
-            row[column.field] === selectedProgram || selectedProgram === ''
-          );
-        }
-        if (column.field === 'last_approval_process_date') {
-          const date = moment(row[column.field]).format('D MMMM YYYY');
-          return date.toLowerCase().includes(searchText.toLowerCase());
-        }
-        return row[column.field]
-          ?.toString()
-          .toLowerCase()
-          .includes(searchText.toLowerCase());
-      }),
-    ) || [];
-
-  const sortedRows = [...filteredRows].sort((a, b) => {
-    if (a[sortField] < b[sortField]) {
-      return sortDirection === 'asc' ? -1 : 1;
-    }
-    if (a[sortField] > b[sortField]) {
-      return sortDirection === 'asc' ? 1 : -1;
-    }
-    return 0;
-  });
+  const filteredRows = filterRows(
+    releasedData?.results || [],
+    'last_approval_process_date',
+    searchText,
+    columns,
+  );
+  const sortedRows = sortRows(filteredRows);
 
   const title = t('Released Payment Plans');
 
@@ -144,40 +109,11 @@ export const ReleasedSection: React.FC<ReleasedSectionProps> = ({
                 {columns.map((column) => (
                   <TableCell key={column.field}>
                     {column.field === 'program' ? (
-                      <Select
-                        value={selectedProgram}
-                        onChange={handleProgramChange}
-                        displayEmpty
-                        fullWidth
-                        size="small"
-                        renderValue={(selected) => selected || 'Programme'}
-                        endAdornment={
-                          selectedProgram !== '' && (
-                            <InputAdornment position="end">
-                              <Box mr={2}>
-                                <IconButton
-                                  size="small"
-                                  onClick={(event) => {
-                                    event.stopPropagation();
-                                    setSelectedProgram('');
-                                  }}
-                                >
-                                  <Close />
-                                </IconButton>
-                              </Box>
-                            </InputAdornment>
-                          )
-                        }
-                      >
-                        <MenuItem value="">
-                          <em>Programme</em>
-                        </MenuItem>
-                        {programs.map((program) => (
-                          <MenuItem value={program} key={program}>
-                            {program}
-                          </MenuItem>
-                        ))}
-                      </Select>
+                      <ProgramSelect
+                        selectedProgram={selectedProgram}
+                        setSelectedProgram={setSelectedProgram}
+                        programs={programs}
+                      />
                     ) : (
                       <TableSortLabel
                         active={sortField === column.field}
